Muldaeng

+++ domain/src/main/java/kr/co/domain/model/SentencesListDto.kt
@@ -3,12 +3,12 @@ package kr.co.domain.model
 data class SentencesListDto(
     var image: String,
     var keyword: String,
-    var keywordId: Int,
+    var keywordId: Long,
     var likeNum: Int,
     var nickname: String,
     var selfLike: Boolean,
     var sentence: String,
-    var sentenceId: Int,
+    var sentenceId: Long,
     var time: String,
     var userId: Long
 )
